requirements.txt
monthdelta
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
braintree==3.39.1
celery==3.1.25
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
confusable-homoglyphs==2.0.2
cryptography==2.1.4
Django==2.0.7
django-activatable-model==1.1.0
django-autoslug==1.9.3
django-cart==1.0.4
django-cities-light==3.5.0
django-entity==3.1.2
django-ipware==1.1.6
django-manager-utils==1.1.1
django-paypal==0.5.0
django-query-builder==1.1.0
django-tastypie==0.14.0
django-user-agents==0.3.2
fleming==0.4.6
idna==2.6
jsonfield
kombu==3.0.37
olefile==0.44
Pillow==4.3.0
progressbar2==3.34.3
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-utils==2.2.0
python3-utils==0.4.0
pytz==2017.3
PyYAML==3.12
realtime-stock==1.0.0
requests==2.18.4
simplejson==3.13.2
six==1.11.0
stripe==1.77.1
suds-jurko==0.6
ua-parser==0.7.3
Unidecode==1.0.22
urllib3==1.22
user-agents==1.1.0
xmltodict==0.11.0
django-payments==0.12.3